+++ src/sim_search_semi_patterns.hpp
@@ -56,7 +56,7 @@ void sim_search_semi_patterns_impl(
   }
 
   if (include_eye)
-    for (int i = 0; i < strings.size(); i++)
+    for (size_t i = 0; i < strings.size(); i++)
       out.insert({i, i});
 }
